Kconfig: Add Application Kconfig with NUM_IRQS
NUM_IRQS is a invisible Kconfig symbol. That means we can't change
with prj.conf or menuconfig.  But we can define yet another NUM_IRQS
in a Kconfig to override it.

Add our own Kconfig for the application with NUM_IRQS in it. we'll
need application configurations in near future anyway.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>