trch_test: Fix port level verification
It used to check the pin level from TRCH side.  But TRCH can't read
the actual signal level but only it's intention, meaning that if it's
driving low, it reads low, regardless of the level at the other end.

Thus, we, the FPGA, must check the level of the pin.  To do so, we use
the test registers.  These registers, together with its IP
cores (GPIOs) allow us to sense the pins driven by TRCH.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>