trch_test: Add test_fpga_pwr_cycle_req
Add a open circuit test for FPGA_PWR_CYCLE_REQ.  This signal is
usually used to inform TRCH to power cycle the FPGA.  But for this
system, the functionality has been removed.  Thus we are free to
change the level of the signal.

The signal is FPGA OUT and TRCH IN.  Thus we drive the signal and
sense the pin on TRCH side.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>