test: Disable UIO3_02 and TRCH_UART_TX while reading ports
I2C_EXT_SCL is connected to either UIO4_08 or UIO3_02.  In case of
UIO4_09, it is connected to the FPGA. Because the FPGA is not powered
on yet, the pin is HiZ. Thus it does not affect this test.  UIO3_02
is, on the other hand, connected to PIC and it's OUT LOW by
default. Thus we make it HiZ for this test.

I2C_EXT_SDA is connected to either UIO4_09 or TRCH_UART_TX. UIO4_09
is, again, connected to FPGA and it's HiZ.  Make TRCH_UART_TX INPUT so
that we can read RD3 safely without TRCH_UART_TX interfering it.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>